requirements.txt
pandas>=1.3.0
psycopg2-binary>=2.9.0
SQLAlchemy>=1.4.0
pathlib
